// Assets/WebGLTemplates/TestTemplate/typescript-tests/UnityInstance.d.ts
export namespace Nahoum.UnityJSInterop.Tests {
export type TestAbstractClass_static = {
fullTypeName_7efcce12970642fa9103e27aca05269f: 'Nahoum.UnityJSInterop.Tests.TestAbstractClass';
assembly_1e2382d1fd5f495db3ec35bf6323452e: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
GetInstance(): TestAbstractClassAbstract;
};

export type TestAbstractClass = {
fullTypeName_87f57789c7da4c3b958c2db16d9fd168: 'Nahoum.UnityJSInterop.Tests.TestAbstractClass';
assembly_1825d2b87e5f4d5294b2173a732aa6dc: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
GetString(): System.String;
}& TestAbstractClassAbstract& TestAbstractClass_static;

export type TestAbstractClassAbstract = {
fullTypeName_9867eee9500541f88cc267c639150c59: 'Nahoum.UnityJSInterop.Tests.TestAbstractClassAbstract';
assembly_1d975a9c9822474fab3b8738a9574e6d: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
GetString(): System.String;
};

export type TestActionCallbacks_static = {
fullTypeName_2177263e8e1244e8a8d4abf42991b183: 'Nahoum.UnityJSInterop.Tests.TestActionCallbacks';
assembly_77dae08f2bfe44a6828a7e754c761334: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
GetInstance(): TestActionCallbacks;
};

export type TestActionCallbacks = {
fullTypeName_c4be2421db5e450db93a170b93a747bc: 'Nahoum.UnityJSInterop.Tests.TestActionCallbacks';
assembly_6db8208171664c3bbc397642c71d18bb: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
TestInvokeCallbackVoid(action: System.Action): void;
TestInvokeCallbackInt(action: System.Action1$Int32): void;
TestInvokeCallbackString(action: System.Action1$String): void;
TestInvokeCallbackFloat(action: System.Action1$Single): void;
TestInvokeDoubleStringCallback(action: System.Action2$Double$String): void;
TestInvokeActionWithClassOutsideNamespace(action: System.Action1$AnotherNamespace_SomeClass): void;
}& TestActionCallbacks_static;

export type TestTasks_static = {
fullTypeName_23c6f3097a02464bae4c65dddc5008ca: 'Nahoum.UnityJSInterop.Tests.TestTasks';
assembly_1219e9a17bd54c00a36c94796f29ee83: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
GetInstance(): TestTasks;
};

export type TestTasks = {
fullTypeName_f6d486e548a040e982d8e2a33a4b604f: 'Nahoum.UnityJSInterop.Tests.TestTasks';
assembly_894e870a1ae2471dbad915999d74dfc4: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
TestTaskVoid(): Promise<void>;
TestTaskString(): Promise<System.String>;
TestTaskInt(): Promise<System.Int32>;
TestTaskFloat(): Promise<System.Single>;
AsyncVoidMethod(): void;
}& TestTasks_static;

export type TestByteArray_static = {
fullTypeName_5449118fd3504fde8a85d5162ba6f132: 'Nahoum.UnityJSInterop.Tests.TestByteArray';
assembly_7ece4577abdf4ef9be712cd6d7dc4b6b: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
LoadTexture(imageBytes: System.Byte_CSharpArray): UnityEngine.Texture2D;
GetTextureResolution(tex: UnityEngine.Texture2D): System.String;
DestroyTexture(tex: UnityEngine.Texture2D): void;
GetByteArrayLength(byteArray: System.Byte_CSharpArray): System.Int32;
};

export type TestExceptions_static = {
fullTypeName_e2a10ec1e5bf4a3da5a8a8f6cb28f6c3: 'Nahoum.UnityJSInterop.Tests.TestExceptions';
assembly_41fe8d721f08418b9b1c7a42a3d26d05: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
ThrowSimpleException(): void;
TestUnraisedException(): void;
ThrowSimpleExceptionAsync(): Promise<void>;
TestUnraisedExceptionAsync(): Promise<void>;
};

export type TestExceptions = {
fullTypeName_5627849c2ff74c879f0378f2ff166204: 'Nahoum.UnityJSInterop.Tests.TestExceptions';
assembly_f6a4eced48f84f669e2ff86630a99506: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
}& TestExceptions_static;

export type TestGenericClass_static = {
fullTypeName_0216b48ba99a4c14ae3f6f2f50f1fbd9: 'Nahoum.UnityJSInterop.Tests.TestGenericClass';
assembly_9c73edf40a944e8b9571b6bf35bec5dc: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
CreateInstance(): TestGenericClass;
};

export type TestGenericClass = {
fullTypeName_612410a61bac4a5fa9000a841592dea5: 'Nahoum.UnityJSInterop.Tests.TestGenericClass';
assembly_53763f59ed644eb6a9c862b90309bbdb: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
GetTypeName(): System.String;
GetTypeNameFromExternal(input: TestInterfaceInGenericClass): System.String;
}& TestInterfaceInGenericClass& TestGenericClass_static;

export type TestInterfaceInGenericClass = {
fullTypeName_3b4d24fc029a4f11ac44692d394d0050: 'Nahoum.UnityJSInterop.Tests.TestInterfaceInGenericClass';
assembly_f0c60cb1751f4cb28c5639e40d2911ae: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
};

export type TestAnotherGenericClass_static = {
fullTypeName_c05d6b15f8094239838eb94f84c5032f: 'Nahoum.UnityJSInterop.Tests.TestAnotherGenericClass';
assembly_7cafe4c602014360953396dc64b0b189: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
CreateInstance(): TestAnotherGenericClass;
};

export type TestAnotherGenericClass = {
fullTypeName_80f6b49c6320400cbed20c26ab76483a: 'Nahoum.UnityJSInterop.Tests.TestAnotherGenericClass';
assembly_4e1eb26bce464844bcf99ba791792127: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
GetTypeName(): System.String;
GetTypeNameFromExternal(input: TestInterfaceInGenericClass): System.String;
}& TestInterfaceInGenericClass& TestAnotherGenericClass_static;
// ...
export type TestGetBasicValues_static = {
fullTypeName_dbdd06c40d244adba3e1a5b5af02da78: 'Nahoum.UnityJSInterop.Tests.TestGetBasicValues';
assembly_aac4224566394f6a85141be795d06d49: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
GetTestString(): System.String;
GetTestDouble(): System.Double;
GetTestInt(): System.Int32;
GetTestBoolTrue(): System.Boolean;
GetTestBoolFalse(): System.Boolean;
GetNullString(): System.String;
};

export type TestGetBasicValues = {
fullTypeName_d38a2014b7f44998bfdd888221b7f5d2: 'Nahoum.UnityJSInterop.Tests.TestGetBasicValues';
assembly_ef2237ee87154f16a8d5deb8e328863b: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
}& TestGetBasicValues_static;

export type TestGetSets_static = {
fullTypeName_52ba95fb545045409192d92cb2c97c2f: 'Nahoum.UnityJSInterop.Tests.TestGetSets';
assembly_e1428a9e24c346278543131465d9f0e7: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
get_TestString(): System.String;
set_TestString(value: System.String): void;
};

export type TestGetSets = {
fullTypeName_c3b827cd75e8466b8557447cd36331c9: 'Nahoum.UnityJSInterop.Tests.TestGetSets';
assembly_51ea787ea3e4432daf4af9753c6a88f1: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
}& TestGetSets_static;

export type TestInstanceMethods_static = {
fullTypeName_43cc6d11113645849c4c68de50085298: 'Nahoum.UnityJSInterop.Tests.TestInstanceMethods';
assembly_fbafde1ce13048148ee3125e6c05e088: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
GetNewInstance(): TestInstanceMethods;
};

export type TestInstanceMethods = {
fullTypeName_31c2f695bcc14c8796a63a27e07d62a4: 'Nahoum.UnityJSInterop.Tests.TestInstanceMethods';
assembly_344d77ebff4a4c05a2cbb8706fb7173e: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
TestGetString(): System.String;
TestGetDouble(): System.Double;
TestGetInt(): System.Int32;
TestGetFloat(): System.Single;
TestGetFloatArray(): System.Single_CSharpArray;
TestGetFloatList(): System.Collections.Generic.List1$System_Single;
TestGetDoubleArray(): System.Double_CSharpArray;
TestGetBoolTrue(): System.Boolean;
TestGetBoolFalse(): System.Boolean;
TestGetNullString(): System.String;
TestAdditionInt(a: System.Int32, b: System.Int32): System.Int32;
TestAdditionFloat(a: System.Single, b: System.Single): System.Single;
TestAdditionDouble(a: System.Double, b: System.Double): System.Double;
TestAdditionString(a: System.String, b: System.String): System.String;
TestGetVector2(): UnityEngine.Vector2;
TestGetVector3(): UnityEngine.Vector3;
TestGetVector4(): UnityEngine.Vector4;
TestGetQuaternion(): UnityEngine.Quaternion;
TestGetColor(): UnityEngine.Color;
TestGetColor32(): UnityEngine.Color32;
}& TestInstanceMethods_static;

export type TestClassImplementingInterface_static = {
fullTypeName_d7936491b6844a369004aa09e2742834: 'Nahoum.UnityJSInterop.Tests.TestClassImplementingInterface';
assembly_2594a0c8b2394ae7bc5ebc06769147fa: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
GetNewInstanceOfInterface(): ITestInterface;
};

export type TestClassImplementingInterface = {
fullTypeName_1f415442d8394fbfa742dd9fd6f3793a: 'Nahoum.UnityJSInterop.Tests.TestClassImplementingInterface';
assembly_ff1016abc615410c9f631cdcb679b044: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
TestGetStringFromClass(): System.String;
TestGetStringFromInterfaceDeclaration(): System.String;
GetSampleFloat(): System.Single;
}& ITestInterface& TestClassImplementingInterface_static;

export type ITestInterface_static = {
fullTypeName_8405f84c959b43d78537628ad86aee39: 'Nahoum.UnityJSInterop.Tests.ITestInterface';
assembly_bdd8140402984ea39bb8e0fa8b30ccfb: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
TestGetStringFromInterfaceStatic(): System.String;
};

export type ITestInterface = {
fullTypeName_853ce4937fea42508a56039cb31b4faa: 'Nahoum.UnityJSInterop.Tests.ITestInterface';
assembly_7c794259fe614e26ace9650518f2910f: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: unknown;
GetSampleFloat(): System.Single;
TestGetStringFromInterface(): System.String;
TestGetStringFromInterfaceDeclaration(): System.String;
}& ITestInterface_static;

export type TestStructExpose_static = {
fullTypeName_f4d743bc1d104fe5a41bb0e85265fdaa: 'Nahoum.UnityJSInterop.Tests.TestStructExpose';
assembly_82657cc113c34b1182471f37b2d6c374: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
GetExposeInstance(): TestStructExpose;
};

export type TestStructExpose = {
fullTypeName_e6a8806ebe0043419429f4a4b7f02fd1: 'Nahoum.UnityJSInterop.Tests.TestStructExpose';
assembly_f055e81d5b8649c994eec8abcce97c45: 'Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null';
value: {A: number, F: number, Name: string};
}& TestStructExpose_static;

}
// ...
export type UnityInstance = {
    Module: {
        static: {
            "Nahoum.UnityJSInterop.Tests": {TestAbstractClass: Nahoum.UnityJSInterop.Tests.TestAbstractClass_static;
TestActionCallbacks: Nahoum.UnityJSInterop.Tests.TestActionCallbacks_static;
TestTasks: Nahoum.UnityJSInterop.Tests.TestTasks_static;
TestByteArray: Nahoum.UnityJSInterop.Tests.TestByteArray_static;
TestExceptions: Nahoum.UnityJSInterop.Tests.TestExceptions_static;
TestGenericClass: Nahoum.UnityJSInterop.Tests.TestGenericClass_static;
TestAnotherGenericClass: Nahoum.UnityJSInterop.Tests.TestAnotherGenericClass_static;
TestGetBasicValues: Nahoum.UnityJSInterop.Tests.TestGetBasicValues_static;
TestGetSets: Nahoum.UnityJSInterop.Tests.TestGetSets_static;
TestInstanceMethods: Nahoum.UnityJSInterop.Tests.TestInstanceMethods_static;
TestClassImplementingInterface: Nahoum.UnityJSInterop.Tests.TestClassImplementingInterface_static;
ITestInterface: Nahoum.UnityJSInterop.Tests.ITestInterface_static;
TestStructExpose: Nahoum.UnityJSInterop.Tests.TestStructExpose_static;
},
        },
        utilities: Utilities,
        extras:{
            "System": {"Action": { createDelegate: (callback: () => void) => System.Action };
"Action<Int32>": { createDelegate: (callback: (a: System.Int32) => void) => System.Action1$Int32 };
"Action<String>": { createDelegate: (callback: (a: System.String) => void) => System.Action1$String };
"Action<Single>": { createDelegate: (callback: (a: System.Single) => void) => System.Action1$Single };
"Action<Double,String>": { createDelegate: (callback: (a: System.Double, b: System.String) => void) => System.Action2$Double$String };
"Action<AnotherNamespace.SomeClass>": { createDelegate: (callback: (a: AnotherNamespace.SomeClass) => void) => System.Action1$AnotherNamespace_SomeClass };
},
        }
    }
}